Move header fetch logic into Page because layouts will not re render even when revalidated
