+++ src/new-client/src/app/lib/components/nhs-page.tsx
@@ -1,5 +1,8 @@
 import Breadcrumbs, { Breadcrumb } from '@components/nhs-breadcrumbs';
 import { ReactNode } from 'react';
+import { NhsHeader } from './nhs-header';
+import { fetchUserProfile } from '@services/appointmentsService';
+import NhsWarning from './nhs-warning';
 
 type Props = {
   title: string;
@@ -8,8 +11,11 @@ type Props = {
 };
 
 const NhsPage = async ({ title, children = null, breadcrumbs }: Props) => {
+  const userProfile = await fetchUserProfile();
+
   return (
     <>
+      <NhsHeader userEmail={userProfile?.emailAddress} />
       <Breadcrumbs breadcrumbs={breadcrumbs} />
       <div className="nhsuk-width-container app-width-container">
         <main
@@ -22,7 +28,16 @@ const NhsPage = async ({ title, children = null, breadcrumbs }: Props) => {
               <div className="app-pane">
                 <div className="app-pane__main-content">
                   <h1 className="app-page-heading">{title}</h1>
-                  {children}
+                  {userProfile === undefined ? (
+                    <NhsWarning title="You cannot access this site">
+                      <p>
+                        You are currently not signed in. To use this site,
+                        please sign in.
+                      </p>
+                    </NhsWarning>
+                  ) : (
+                    children
+                  )}
                 </div>
               </div>
             </div>
