+++ src/new-client/src/app/page.tsx
@@ -4,13 +4,10 @@ import NhsPage from '@components/nhs-page';
 
 const Home = async () => {
   const userProfile = await fetchUserProfile();
-  if (userProfile === undefined) {
-    return null;
-  }
 
   return (
     <NhsPage title="Appointment Management Service">
-      <SiteList sites={userProfile.availableSites} />
+      <SiteList sites={userProfile?.availableSites ?? []} />
     </NhsPage>
   );
 };
